port idents to wacc

+++ include/frontend/intermediate/idents.h
@@ -17,7 +17,7 @@ TIdentifier rslv_var_identifier(struct IdentifierContext* ctx, TIdentifier varia
 TIdentifier rslv_struct_tag(struct IdentifierContext* ctx, TIdentifier structure);
 TIdentifier repr_label_identifier(struct IdentifierContext* ctx, LABEL_KIND label_kind);
 TIdentifier repr_loop_identifier(struct IdentifierContext* ctx, LABEL_KIND label_kind, TIdentifier target);
-TIdentifier repr_case_identifier(struct IdentifierContext* ctx, TIdentifier target, bool is_label, size_t i);
+TIdentifier repr_case_identifier(struct IdentifierContext* ctx, TIdentifier target, bool is_label, unsigned long i);
 TIdentifier repr_var_identifier(struct IdentifierContext* ctx, struct CExp* node);
 
 #endif
